Basic annotations - floor

+++ 0x00-python_variable_annotations/2-floor.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
-import math
 """Basic annotations - floor"""
+import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """Returns the floor of n"""
     return math.floor(n)
